src/main/java/issuesandalgorithms/SortingAndSearching.java
package issuesandalgorithms;

public class SortingAndSearching {
    //11.1
    public static void merge(int[] a, int[] b, int lastA, int lastB) {
        int indexMerged = lastB + lastA - 1;
        int indexA = lastA - 1;
        int indexB = lastB - 1;

        while (indexB >= 0) {
            if (indexA >= 0 && a[indexA] > b[indexB]) {
                a[indexMerged] = a[indexA];
                indexA--;
            } else {
                a[indexMerged] = b[indexB];
                indexB--;
            }
            indexMerged--;
        }
    }

}
